+++ purrnet/Assets/PurrNet/CHANGELOG.md
@@ -1,3 +1,25 @@
+# [1.23.0-beta.23](https://github.com/PurrNet/PurrNet/compare/v1.23.0-beta.22...v1.23.0-beta.23) (2026-09-03)
+
+
+### Bug Fixes
+
+* frame-accurate RTT; immediate rpcs also send on the same frame ([6ab4fad](https://github.com/PurrNet/PurrNet/commit/6ab4fadd785d2c3f5bd5f70c9b3506623d7fd400))
+
+# [1.23.0-beta.22](https://github.com/PurrNet/PurrNet/compare/v1.23.0-beta.21...v1.23.0-beta.22) (2026-09-03)
+
+
+### Bug Fixes
+
+* ILProcessing for unity 6.6 ([bdbbb82](https://github.com/PurrNet/PurrNet/commit/bdbbb8205c66f8729f707892e201f582a2c74824))
+* when using Purrnet, Object.DontDestroyOnLoad() can fail ([2370872](https://github.com/PurrNet/PurrNet/commit/237087254d76d0cafa9f55c40ce1d2368caae0be))
+
+# [1.23.0-beta.21](https://github.com/PurrNet/PurrNet/compare/v1.23.0-beta.20...v1.23.0-beta.21) (2026-09-03)
+
+
+### Bug Fixes
+
+* PlayerSpawner not spawning if DespawnIfOwnerDisconnects = false && ([62b2a7f](https://github.com/PurrNet/PurrNet/commit/62b2a7f7f32bb15580c418a14df2c71134e8945c))
+
 # [1.23.0-beta.20](https://github.com/PurrNet/PurrNet/compare/v1.23.0-beta.19...v1.23.0-beta.20) (2026-09-03)